02-cloud-function/requirements.txt
flask==2.0.3
werkzeug==2.0.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0